% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignDepth.R
\name{KSign}
\alias{KSign}
\title{\code{K}-sign depth for univariate data}
\usage{
KSign(x, K, naive = FALSE)
}
\arguments{
\item{x}{A vector of -1 and +1 of length \code{n}. These
are typically the signs of the residuals (-1 for negative sign,
+1 for positive sign).}

\item{K}{Positive integer, the parameter of the depth.}

\item{naive}{Logical indicator whether the naive implementation
in R should be used. By default set to \code{FALSE}. Use \code{FALSE}
in all situations, the naive regime is only for checking and comparison.}
}
\value{
If \code{naive=FALSE}:
A vector of length \code{2K} of number of occurences
of alternating patterns in the sequence \code{x}. The vector
is named \code{+j} or \code{-j}, where the initial sign
stands for the starting sign of the alternating sequence
(-1 for - and +1 for +), and \code{j} is the length of the sequence.
\code{+3} thus means the number of triples of the form \code{+-+}
in the sequence \code{x} etc. The final \code{K}-sign depth is
the sum of the final two entries of this vector (\code{+K} and \code{-K})
divided by \code{choose(n,K)}, where \code{n} is the length of \code{x}.
The function also provides all \code{j}-sign depths for \code{j<K};
one just needs to sum the two corresponding entries in the vector.

If \code{naive=TRUE}: Total number of \code{K}-sign alternating
sequences. This corresponds to the sum of the two elements in the
last row of the output matrix if \code{naive=FALSE}.
}
\description{
C++ and R implementations of the algorithm for
the computation of the \code{K}-sign depth for
one-dimensional data (residuals). The C++ implementation
is the fastest one that should be used in practice; its
complexity is O(n). For comparison,
also the (very slow) naive R version of complexity O(n^K)
is provided.
}
\examples{
n = 500
x = sign(rnorm(n))
K = 3
sum(KSign(x,K)[K,])/choose(n,K) # K-sign depth

n = 50
x = sign(rnorm(n))
K = 3
sum(KSign(x,K)[K,])==KSign(x,K,naive=TRUE)
}
